Refactor summaryService.js: extract generation and session-complete logic
Extract core generation engine and session-completion lifecycle helpers into
dedicated modules (summaryGenerate.js and summarySessionComplete.js) to keep
the main orchestration module under the line limit while maintaining separation
of concerns. This improves code maintainability without changing functionality.

- summaryService.js: Keep orchestration logic, import refactored helpers
- summaryGenerate.js: Core generation engine with private helper functions
- summarySessionComplete.js: CI-check scheduling and outcome mapping
- ConversationPanel.vue: Minor formatting improvement
